\name{multisphere-package}
\alias{multisphere-package}
\alias{multisphere}
\docType{package}
\title{\packageTitle{multisphere}}
\description{\packageDescription{multisphere}}
\section{Package Content}{\packageIndices{multisphere}}
\author{\packageAuthor{multisphere}}
\section{Maintainer}{\packageMaintainer{multisphere}}
\keyword{package}
